requirements.txt
autodistill
autodistill_clip
autodistill_grounded_sam
numpy
opencv-python
pillow
roboflow
supervision
tqdm
pycocotools
scikit-image
torch
torchvision
colorlog